requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==1.1.2

# Flask Packages
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-BabelEx==0.9.4
Flask-User==1.0.2.2
Flask-WTF==0.14.3

# Automated tests
pytest==5.4.1
pytest-cov==2.8.1

# Libraries
beaker==1.11.0
boto3==1.12.39
blinker==1.4
celery==4.4.2
click==7.1.1
ldap3==2.7
psycopg2-binary==2.8.5
pyyaml==5.3.1
requests==2.23.0
email_validator
gpiozero

# Development tools
# tox==3.5.2
